Moved web api generators for .net framework and .net core to own projects
