+++ src/NetTypeS/NetTypeS.WebApi.Framework/packages.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.6" targetFramework="net462" />
+  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.6" targetFramework="net462" />
+  <package id="Newtonsoft.Json" version="6.0.4" targetFramework="net462" />
+</packages>
