+++ src/NetTypeS/NetTypeS.sln
@@ -11,11 +11,15 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Examples", "Examples", "{E2
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "FrameworkExamples", "FrameworkExamples", "{249FC396-24EE-4A3F-BE66-6CB7782DCF65}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetTypeS.ExampleData", "NetTypeS.ExampleData\NetTypeS.ExampleData.csproj", "{D3287629-9CAB-4F19-9827-2C05F3CF43E8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetTypeS.ExampleData", "NetTypeS.ExampleData\NetTypeS.ExampleData.csproj", "{D3287629-9CAB-4F19-9827-2C05F3CF43E8}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetTypeS.FrameworkWebExample", "NetTypeS.FrameworkWebExample\NetTypeS.FrameworkWebExample.csproj", "{6B16EC7C-8D0B-4144-8704-24FFF7BF2704}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetTypeS.FrameworkExample", "NetTypeS.FrameworkExample\NetTypeS.FrameworkExample.csproj", "{F7A375C9-6CBB-40FD-A34B-88290E0CE2A7}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetTypeS.WebApi.Framework", "NetTypeS.WebApi.Framework\NetTypeS.WebApi.Framework.csproj", "{E5257A29-1735-4CEA-BB3A-6B0FF567C020}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetTypeS.WebApi.Core", "NetTypeS.WebApi.Core\NetTypeS.WebApi.Core.csproj", "{4DF2A523-339E-4F77-8269-233851F515A4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -43,6 +47,14 @@ Global
 		{F7A375C9-6CBB-40FD-A34B-88290E0CE2A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F7A375C9-6CBB-40FD-A34B-88290E0CE2A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F7A375C9-6CBB-40FD-A34B-88290E0CE2A7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E5257A29-1735-4CEA-BB3A-6B0FF567C020}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E5257A29-1735-4CEA-BB3A-6B0FF567C020}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E5257A29-1735-4CEA-BB3A-6B0FF567C020}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E5257A29-1735-4CEA-BB3A-6B0FF567C020}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4DF2A523-339E-4F77-8269-233851F515A4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4DF2A523-339E-4F77-8269-233851F515A4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4DF2A523-339E-4F77-8269-233851F515A4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4DF2A523-339E-4F77-8269-233851F515A4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
